requirements copy.txt
certifi==2018.8.24
cffi==1.11.5
click==7.1.1
cycler==0.10.0
Flask==1.1.1
gevent==1.3.6
greenlet==0.4.13
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.0
matplotlib==3.0.3
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.32
Pillow==5.3.0
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
six==1.14.0
torch==1.2.0
torchvision==0.4.1
Werkzeug==1.0.0
wincertstore==0.2
